requirements-server.txt
flask
flask-cors
gunicorn
anthropic
google-genai
